refactor(tests): split autocomplete context_tests into focused modules
Split the 1162-line context_tests.rs into multiple focused test modules:
- common.rs: Shared test helpers (tracker_for, create_array_of_objects_json)
- basic_context_tests.rs: Basic context detection tests
- char_before_field_tests.rs: Tests for find_char_before_field_access helper
- object_key_context_tests.rs: Object key context detection tests
- property_tests.rs: Property-based tests using proptest
- element_context_tests.rs: Element context integration tests

Follows project conventions from insertion_tests/ structure.
All 59 tests passing.
